Add support for event hooks
Although the standard means of retrieving events is sufficient for many
cases, there are scenarios where out-of-band snooping of events is
required. Consider the case where a widget to reset part of its own
state whenever a key is pressed, while there is no requirement that said
widget have the input focus.
This change implements such an out-of-band event retrieval mechanism in
the form of event hooks. Each widget can register an event hook with the
UI and subsequently will receive every event before any actual widget
gets to handle it.