Return targeted event when no receiver is present
"Ordinary" events such as key presses are delivered directly to the
focused widget. However, it is possible for no widget to currently have
the input focus. Currently we simply return None in such a case, but
that is the opposite of what we do when an event is not handled by a
widget, where we pass it through.
This change aligns the two cases by properly returning the event in case
no widget has the input focus and, hence, can handle it.