Refactor event handling internals some more
The event handling code is quite complex and with that hard to
understand and reason about. In order to make it easier digestible, we
try to make the various functions involved tackle different concerns of
the handling of events.
This change is another step into this direction. It introduces the
following separation of responsibilities:
- handle_meta_event merely recurses into chained events
- handle_ui_event only dispatches a UiEvent object
- handle_event invokes the widget event handler and determines the
  parent which will handle the returned event, if any