Make sure to always chain resulting events together
In the past we have made the decision to that although widget event
handlers may return multiple events (in chained form), only a single
event is allowed to be returned from the Ui object's handle method.
This stance is being challenged with upcoming use cases.
This change prepares for a change of the return value of said handler.
In particular, it makes sure to create an accumulated chain event from
all the individual events returned from widget handlers as opposed to
dropping every but the last (we still drop every event but the last
currently, so no change in semantics should be perceived).