Allow for event hook functions to return events
So far we have allowed the recently introduced event hook functions to
only inspect events and potentially alter widget state based on that.
In some cases, though, we would like for event hooks to also be able to
emit new events.
To accommodate such cases, this change adjusts the signature of the
event hook function to include the return of an optional MetaEvent
object, just as normal event handler methods do. Events returned this
way will be chained together with the actual event being handled and be
go through the normal handling pipeline as such a chain.