Make Ui::handle return an optional MetaEvent
This change adjusts the return value of the Ui's handle method to no
longer return an UiEvent but a MetaEvent. The MetaEvent struct was
designed mainly as a means to facilitate communication between widgets
in the UI. Hence, at that time there was no need to return such events
(which can comprise multiple UiEvent objects) from the Ui to the
outside. Instead, we merely returned the last event in the chain. While
that was a valid stance, the silent dropping of events has always been
regarded with skepticism.
This patch corrects this problem and introduces more transparency by
returning all unhandled events in the form of an event chain. It should
be seen as a change to provide client code with better insight into the
event handling aspects of a particular UI.