Remove functionality for retrieving the last focused widget
This change removes the functionality for retrieving the last focused
widget (and all the code used for maintaining the value of the
corresponding internal variable). The reason for this removal is two
fold.
First, after having gotten more experience with UI workflows with the
crate, the believe is that remembering the last focused event on the
client side may be the more appropriate solution after all and that with
more complex UIs it may indeed be required to funnel events that would
require knowledge of the last focused widget through a widget that would
be equipped to keep track of this information.
Secondly, storing the last focused widget has always been regarded as a
policy decision. That is problematic, because some clients may not use
this functionality at all (in which case resources are wasted on it),
while others may require an entire stack of previously focused widgets
-- a requirement that we cannot and would not want to fulfill.