Do not focus any widget by default
Among the first (minor) decisions was that a newly created widget should
become the focused one if no other has the current input focus. This
functionality has always been regarded as opinionated, while at the same
time making a reasonable amount of sense.
Be that as it may, this patch removes this conditional logic. There are
a couple of reasons leading to this decision. First, it is a policy and
as such does not really belong into the core of the UI. Second, cases
are imaginable where this behavior is actually detrimental to
correctness and the user -- should he/she be aware of this behavior --
would have to work against it (if there is no awareness then this sort
of behavior is the source of a bug). Last, and most importantly, it
complicates the internal logic. Right now that complication is pretty
much limited to the couple of lines comprising the logic itself, but in
the future when we change the order of children in the parent (i.e., the
z-order) when changing visibility/focus this complexity would lead to
special cases that have the potential to significantly complicate the
logic.
By removing these few lines we make the code more future proof to
upcoming additions.