Reorder child order on visibility operations
So far the z-order of a widget, i.e., its position in the vector of
children its parent maintains, has been determined only by the order in
which it was created and would not change. However, when thinking of a
UI and a widget receiving the input focus, it and its parents would
subsequently appear raised above all other widgets.
This change implements such logic for all widgets managed by this crate.
All three visibility changing operations (focus, show, hide) now also
adjust a widget's order in its parent's list of children. Strictly
speaking, only focusing operations are required to trigger a reordering
of that sort, but we opted to implement this logic holistically. This
way, if a widget is hidden it will be moved all the way back before the
first previously hidden widget. Similarly, a previously hidden widget
that is made visible, will be moved up just past the last widget that is
already visible.