Pass Cap reference to rendering infrastructure
Depending on the type of UI a renderer may need to highlight focused
widgets specially. Currently we do not support this case, as there is no
way to retrieve this state from the widget itself which gets passed to
the renderer object.
This change adjusts the code to pass a Cap reference to the renderer
method, in addition to the actual widget and the bounding box to use.