Bump version to 0.1.0
This change bumps the version of the crate to 0.1.0, which marks the
first stable release. The crate should no longer be considered to be in
an early development phase and completely experimental. Notnow, a
terminal based task management application, is successfully developed
using event dispatching and widget management functionality provided by
the crate.
Along with that version bump, this change updates the README file to
reflect the current status more appropriately.