Separate out custom event handling infrastructure
So far we have treated custom events similar to "normal" events. That
is, they all were represented as variants in the Event type. This design
was identified to be incompatible with a future direction we want to
head into: we would like to support custom events that are guaranteed to
be returned to the originating widget after they got handled. As part of
this handling they may get modified. Ultimately what this means is that
instead of transferring ownership of an event to a handler function,
only a mutable reference would be passed. Subsequently, the UI proper
would make sure to deliver the potentially modified event back to the
caller.
Long story short, the current event handling infrastructure needs to be
reworked slightly to enable this use case. In particular, this change
separates the handler methods that widgets have to provide for "normal"
events (currently only key presses/releases) and custom events (i.e.,
boxed up Any objects).
The reason we need this separation is because we want only custom events
to be returnable in this fashion. However, if we were to pass a mutable
reference to an Event that includes a custom event to an event handler,
there is no guarantee that said event will still be of the Custom
variant that it was beforehand. The widget could have changed it to be a
key event now.
By having a separation between custom events and ordinary ones we
circumvent this problem.
A welcome side effect of this change is that by removing the Custom
variant from the Event enum, said enum now implements Send and can be
send over thread boundaries.