Implement Copy, Clone, and PartialEq traits for Event type
With the recent removal of the Custom variant from the Event type, we
gained the ability to implement various traits on said type, that
previously could not be implemented due to restrictions imposed by the
contained Box<Any> object.
To that end, this change implements the Copy, Clone, and PartialEq
traits for the Event struct. Doing so simplifies the test code a bit.