Adjust event hook function signature to take event by value
With the recent changes making Event a light weight and, in fact,
copyable type, we can look where it may make more sense to pass in
events by value as opposed to by reference. On such location is the
event hook function.
This change adjusts the signature of the event hook function to take the
event by value as opposed to by reference.