Introduce "returnable" custom events
Under some circumstances widgets may want to send an event to another
widget and have a guarantee that the content of said event will be sent
back to them -- in the same or modified form. Although support for such
use cases can already be built with the existing infrastructure, it
seems advantageous to implement such functionality in the crate itself
in order to have only a single implementation and to reduce code
duplication and custom logic in client code.
To that end, this change introduces another custom event variant to the
UiWidget type: Returnable. This variant encodes the aforementioned
desire of the sending widget to receive back the event after it was
shown to a widget. A Ui object will ensure that this invariant is
enforced.
In order to implement this scenario fully, widgets get yet another event
handling method that accepts a mutable reference to a custom event (that
is, to an Any object). An alternative implementation could have shared
the existing custom event handler and provided a custom event (in the
form of an enum) that can contain either a borrowed or an owned event
(similar to the internally used CustomEvent). However, given that we
foresee custom events to be typically sent as either directed or
returnable ones, a separation of the two handlers seems like the more
appropriate design.