Rename UiEvent variant of MetaEvent to Event
The MetaEvent enum got introduced as a way to create a chain of UiEvent
objects that can be processed recursively. In the future we want to move
away from the strict focus on UiEvent and instead allow for chaining of
arbitrary event types.
This change marks a first step into this direction by renaming the
'UiEvent' variant of MetaEvent to the more generic 'Event'.