Introduce ChainEvent struct
An upcoming change will introduce the means to chain arbitrary events
together. To pave the way for this design, this change introduces a
new struct, ChainEvent, that, in its current form, is just a different
name for the MetaEvent enum. In a next step this newly introduced enum
will actually be made parameterizable by the event type to use.