Move event related test functionality into only module using it
This change moves the compare and clone functionality for the various
event types into the only module that is still using them: the test
module for events.
This functionality was never considered to be well designed, but for
lack of a better alternative it was introduced nevertheless. However,
given this stance it is prudent to prevent it from spreading into other
modules. Because only a single module is actually using those functions,
moving them there provides for much better isolation and will prevent
"accidental" usage in other modules in the future.