Remove event related clone methods
As it turns out, with a bit of refactoring we can do away with the
internally used clone functions for the UiEvent and MetaEvent types.
These functions, along with their *compare counterparts, were introduced
to simplify testing, but they are ultimately not very Rusty in style and
their usage should be minimized.
Removing the clone functionality is the right step in this direction.
Unfortunately, no way has been found to remove the *compare functions as
well without blowing up the amount of code significantly.