Introduce new UnhandledEvent type
So far the UI proper has treated input and output events in exactly the
same. That is, they were literally of the same type: MetaEvent, which is
a chain of UiEvent objects.
However, by now the UiEvent enum has two variants that merely define the
addressing for custom events. Those variants make sense in the input
context, but returning them from the Ui object's handle method is
arguably wrong, as a caller can never receive a Directed or Returnable
custom event.
This change cements this assurance by means of the type system. It
introduces a new type, UnhandleEvent, that comprises the non-addressing
related variants of the UiEvent. Furthermore, a chain of those events is
now returned from the Ui's handle method.