Rename MetaEvent to UiEvents
In a struggle to find appropriate names for types in general and events
enums in particular, this change performs another renaming. The by now
well-kown MetaEvent is renamed to UiEvents. The rename has three
reasons:
1) MetaEvent was never found to be a perfect name, but proper naming is
   arguably hard and we had to roll with something.
2) UnhandledEvent has a chained variant that is referred to as
   UnhandledEvents.
3) The focus of a ChainEvent (or MetaEvent in particular) has always
   been to combined multiple events (UiEvents in the case of MetaEvent)
   and not on actually introducing a distinct new event type to reason
   about.

With this change we are arguably moving a step closer to just returning
a vector of events -- and this possibility has certainly been considered
-- but we still see value in having an Option<UiEvents> as opposed to a
Vec<UiEvent> returned, although most of this preference may not be on a
strictly technical level.