Enable Rust 2018 and future compatible lints
This change enables the rust-2018-compatibility and rust-2018-idioms
lints. These lints subsume the previous enabled bare-trait-objects and
elided-lifetimes-in-paths ones, which consequentially get removed. We
furthermore enable the future-incompatible lint which hopefully catches
other potential problems early on.
In order to enable those lints we have to fix up a couple of pub items
that are unreachable from the crate root, by making them crate local,
and replace module paths to use the new 'crate::' syntax.