Avoid unnecessary clone when invoking event hooks
When invoking the event hooks for the widgets that registered to receive
them we have to clone the hash set we use for keeping track of who
registered a hook to begin with. The reason for this peculiarity is that
a mutable borrow of the Ui is required to invoke the hook and so we
cannot hold an immutable borrow to part of it by virtue of having an
iterator over a member.
This change fixes this problem. We sidestep the clone by working with
indices instead of using an iterator. In order to be able to do that we
replace the HashSet with a Vec. We keep the vector sorted and use binary
search to insert and lookup elements in it.