Deliver event hook emitted events after the source event
Event hooks are allowed to emit events on their own (but they cannot
modify the source event). So far, these newly emitted events were
delivered to the widget before the original source event (but generally
this was an unspecified property; the order of event delivery was left
open).
With this patch we change the order of delivery. From now on events
emitted from a hook will only reach the handling widget after the source
events themselves did.
The reason for this change in behavior is that it can allow for some
nice event handling patterns to be built, because it allows for event
reordering to happen. For example, consider the case where a widget
hooks all incoming events and, if an event is of a certain type (say, a
key press), it wants to clear some internal state. While entirely
possible already, problems may arise in case the event being hooked will
actually also be handled by the hooking widget. In such a case internal
state may first be cleared and then be set again. Depending on the use
case that can be a problem.
By having the property that hook emitted events are delivered after the
source event itself, we can essentially schedule a guaranteed "call
back" by emitting an event from the event hook. Once that event is
received by the widget, it is guaranteed that the source event had been
handled and internal state could have been updated. The clearing in the
above example can now happen taking into consideration whether the
source event modified internal state as well.