Add test for OptionChain::opt_chain method
When we introduced the OptionChain trait we forgot to add a test that
explicitly tests the its opt_chain method. Although implemented in terms
of other tested chaining functionality, that may not always be the case
and we should have a test irrespective of that. This change adds such a
test.