Add badges for crates.io, docs.rs, and minimum rustc version
This change adds another three badges to the repository. The first two
simply list the most recent version of the crate as published on
crates.io and of the docs as published on docs.rs, respectively. The
third one states the minimum version of rustc that is required for
building. We have recently switched to using Rust 2018 and with that we
have a requirement for version 1.31 of the toolchain. In the future the
hope is that now that we have a proper CI/CD pipeline based off of
Docker images we should be able to build on past versions of Rust, even
if they are not used during development.