Introduce 'derive' feature re-exporting custom derive functionality
Conceptually the custom derive functionality provided by the gui-derive
crate belongs into the gui crate itself. The reason it cannot reside
there is rooted in Rust peculiarities that could change at some point
and, ultimately, is an implementation detail.
To abstract away from this detail, this change re-exports this custom
derive functionality from the gui crate itself, making it more
conveniently available to users. This export is guarded by a feature
switch (which is enabled by default), to be able to opt-out from relying
on this dependency.