Drop "Gui" prefix from procedural macros
Up until Rust 2018, macros shared a global namespace. That is, an
exported macro had the potential to collide with a macro of the same
name exported by another crate. To circumvent those problems, prefixes
were typically used.
However, since Rust 2018 this behavior changed and macros can be
imported on demand and on a per-crate basis. Hence, the need for
prefixes vanished.
To that end, with this change we drop the "Gui" prefix that we used for
the custom derive attributes, making the names more concise.