Represent New enum as Option<()>
In the past we have used a dedicated enum to represent the
gui(default_new) attribute internally. That is a bit cumbersome and in
fact quite clumsy.
With this change we represent this piece of information simply with an
Option<()> (although in the form of a typedef), making the code more
concise.