Factor out expand_handleable_trait function
We have a bit of an asymmetry in the expansion structure: for the Widget
expansion we have a expand_widget_input function that merely checks that
we are working with the proper type as well as a expand_widget_trait
one, that does the bulk of the work.
For the Handleable trait we only have a expand_handleable_input function
that does all the work.
With this change we introduce a new function, expand_handleable_trait,
similarly to what we have for handling the Widget trait.