Add TypeId functionality to Renderable
Some time back we introduced TypeId functionality for Any-style casting
to the Widget. That allowed us to avoid the direct usage of Any, which
is basically implemented by any arbitrary type. Instead, an actual
Widget had to be supplied, preventing accidental usage of types not
implementing this trait.
With this change we provide similar functionality for the Renderable
trait. Renderable is one of the consumers of this logic (the other
effectively being client code, for example, when invoked by event
hooks).
Adding this functionality to Renderable helps us isolate the individual
traits from each other. Currently, a Renderer works with a Widget, but
really it should only be a Renderable. Client code should cast into an
actual widget if required.
Doing so will also be a requirement in the future when we make Widget
generic over the event type used, because we do not want this event type
to permeate into the rendering infrastructure.