Adjust Renderer::render to work with a Renderable
This change adjusts the Renderer's render method to work with a
Renderable instead of a Widget. This step helps in better decoupling the
individual traits from each other, but it also is a requirement for
future changes that parametrize the Widget by the event type used, such
that the Renderer can continue to be agnostic of the widget's event
type.