Adjust event hook function signature to take event by reference
In the past we deliberately changed the signature of the event hook
function from one that takes the event via reference to one that takes
it by value. This change was motivated by the fact that Event
implemented Copy.
In the future we want to support arbitrary event types and we do not
want to enforce that those implement Copy. Hence, with this change we
adjust the event hook function signature back to one taking the event by
reference.