Make UiEvent & UnhandledEvent generic over the wrapped event type
In the future we want the crate to no longer force users to work with a
specific event type. Instead, they should be free to use whatever event
type they have at hand.
As a first step towards making this vision reality, this change adjusts
the UiEvent and UnhandledEvent types to be generic over the internally
used event. Note that ChainEvent has been generic all along, because
it was required to work with more than one event type.