Introduce MutCap trait
This change splits the existing Cap trait in two: Cap and MutCap. The
Cap trait can be used for all operations that only read from a widget.
MutCap extends Cap with objects that require a mutable self. Right now
this split provides no real value -- we already supported exactly these
semantics just by virtue of working with an immutable or mutable
reference to the original Cap. However, we need this separation for a
slightly more subtle reason: in the near future a Widget (and a
Handleable) will be parametrized by the event type to use. Because Cap
(the original version of it) contains an add_widget method that works
with a Widget, Cap would now have to be parametrized by the event type
as well. That in turns leads to a leak of the event type into the
Renderer (which uses a Cap as well, but only an immutable one), which is
undesired.
Because an immutable Cap could never be used to add a widget to the UI,
this leakage is unnecessary and can be prevented by splitting off the
add_widget method. The split we decided to perform with this change is
to simply group operations based on the mutability they require. Now the
only entity using a (parametrized) MutCap is Handleable, and it is
parametrized by the event to use as well.