Add support for gui(Event = ...) attribute
In a little while the Widget and Handleable will become generic over the
event type used. That puts the custom derive functionality into somewhat
of an unfortunate spot: we want to support both widgets that are
implemented for only a particular type of event as well as those that
are completely generic over the event type.
To solve this problem, this change adds support for a new attribute:
gui(Event = ...). This attribute, if present, will determine the event
type to specialize the widget for. If it is not specified by the user,
the implementation will be completely generic.
Note that as of now, while all the plumbing is done to evaluate and loop
through this event type, code emission is still left untouched.