Fix up From implementations for the various event types
We provide a bunch of convenience conversion between the various event
types. For example, a single event can be converted into a chain of
UiEvents. The problem with the current implementation of those
conversions is that they are too generic. The conversion from a single
event into E an UiEvents object is conceptually fine, but the definition
requires that E: Into<UiEvent<Event>>. Once we go fully generic, this
multi-level genericity means that we have to carry around those
additional constraints everywhere.
With this change we rework the conversions a bit so as to be only
generic over a single parameter.