Make Handleable generic over the event type to use
The crate aims to be as general as possible when it comes to development
of a UI. A severe violation of this striven for generality is the fact
that it provides its own Event class, effectively dictating clients what
an event has to look like.
This change addresses this problem by making the Handleable trait (and
with it the Widget trait) generic over the event type to use.