Use custom event for tests
Now that all widgets can be parametrized by the event to use there is no
more reason for the crate to actually contain an event type.
In preparation for removing this type, this change converts the tests to
use a newly introduced custom event type, suitable only for unit testing
functionality.