Correctly handle generics in custom derive functionality
When the Widget trait was made generic over the event type to use we
initially -- for the sake of simplicity -- assumed a concrete event type
to be used together with the custom derive macros. This is the most
common use case, as widgets typically rely on a concrete type in order
to implement their behavior.
However, two cases were unaccounted for: First, a widget itself may be
parametrized by a type, which we should be handling transparently. And
second, the event type provided should be allowed to be generic. For
example, it is possible to envision cases where a user has different
event types that implement a common trait.
This change adds support for both cases to the custom derive macros we
implement.