Add missing 'dyn' keyword for tests
When we got the crate ready for Rust 2018 we added a bunch of lints to
the library and, among other things, added the 'dyn' keyword to trait
object uses. However, the integration style tests act as separate crates
and were not affected.
With Rust 1.37 the usage of trait objects without the 'dyn' keyword has
officially been deprecated and is causing a warning to be emitted. With
this change we adjusts these tests accordingly.