Enable missing docs lint
Due to a bug in the Rust compiler which caused it to incorrectly flag
crate level documentation as missing in a proc-macro crate even if it
really existed, we were unable to enable this lint for the gui-derive
crate.
However, with Rust 1.36 this problem has been resolved and so with this
change we enable the lint.