Bump minimum Rust version to 1.36
As part of an upcoming change we want to enable the missing-docs lint
for the gui-derive crate. This has not been possible due to a bug in the
Rust toolchain, manifesting as a wrongly emitted warning/error, but with
Rust 1.36 this problem has been resolved.
To that end, this change bumps the minimum required version of Rust
accordingly.