Switch to using master as the main branch
This change switches to using the master branch as the main branch, as
is usually the case for git. In the past we used the devel branch as the
development branch on which rebases are performed relatively freely, so
as to incorporate small fixups.
While this may still be the case, we now feed CI/CD badges from the
master branch and will "merge" into master with every new release.