Add link to coverage run information
As it turns out GitLab actually provides a static address that
dynamically resolves to the artifacts of the most recent CI run. Given
that our code coverage visualization is just an artifact in the form of
an HTML page, we can actually link to it directly. This change does
exactly that.