Bump minimum Rust version to 1.42
As part of an upcoming change we want to make use of a feature that only
got introduced recently, with Rust 1.42.
To that end, this change bumps the minimum required version of Rust
accordingly. In order to go about that it also fixes a few new warnings
emitted by clippy.