Use FnOnce for widget creation
In the past we have reluctantly used an FnMut based signature for the
widget creation function. That can lead to cumbersome code, as it may
mean doing some unwrap on an Option to cater to cases where a client
wants to move something into the function.
With this change we use a true FnOnce for this purpose, which eliminates
this issue. It also introduces a new one, in that we now have to box up
this function everywhere, because closures are unsized and so one cannot
simply move out of them. This, however, should get addressed once the
unboxed_closures feature stabilizes.