Reorder Cap & MutCap parameters
In the past we have mostly kept the Cap/MutCap parameters as the last
ones in any function/method signature, because we saw the provided
references as auxiliary and the other arguments as of greater
importance.
With this patch we change this stance: this parameter is now meant to be
seen as an extension of the widget and should follow it. That is, in all
methods of the Handleable trait, it is now on second position after
self. In functions where a widget ID is present we order it after that.