Separate widgets and their mutable data and store them separately
One of the long standing problems in our design has been the fact that
we "take" widgets out of the Ui object's data structures to work around
the fact that we need a mutable Widget reference when handling an event,
but we also want to provide a MutCap, which means a mutable Ui object
reference, to provide to the widget being called.
The fundamental underlying problem can be regarded as us requiring a
mutable widget to begin with. To resolve this problem, this change
proposes to separate the Widget object from its associated data. While
the latter in all likelihood needs to be mutable, the former really
does not.
To that end, we now require two functions when creating a widget: one
for creating the data associated with it and another for creating the
widget itself. We furthermore adjust our Cap and MutCap traits to look
up the data associated with a widget, by providing its Id.