Introduce short hand data accessor methods
With the earlier change separating the Widget object from its data,
clients need a lot of boilerplate for accessing this data, involving
looking up and providing the widget's Id, retrieval of the data, and a
cast to the expected type.
To shorten this sequence, this change introduces two new methods to the
Widget trait that perform this job automatically. We introduce these in
the Widget trait and not Object (which has the id() method that we use)
for two reasons:
1) Object is more or less an implementation detail and
2) Object does not know our event type E, so we need an additional
   generic parameter in the data_mut method call there for MutCap.

To make the new data function also work on a MutCap reference, we now
require MutCap to Deref to Cap.