Use non-mutable target in event hook functions
Similar to the methods of the Handleable trait, the event hooks no
longer require a mutable Widget reference, because they can simply
retrieve a mutable reference to the widget's data through the provided
capability object and work on it.
To that end, this change adjusts the signature of the event hook
function to provide a non-mutable Widget reference.