Simplify widget reordering code
Back in the days when one of the first versions of the crate was
created, the compiler did not yet have support for non-lexical lifetimes
(NLL) and we had to do some acrobatics to make certain code pieces work.
By now NLL has been available for quite a while and with this change we
make better use of them, simplifying the code in question a bit.