Store widgets behind an Rc
With Widget objects no longer requiring any modification by clients and
us no longer handing out mutable references to them, we can now turn to
removing the "taking" of widgets from the Ui for Handleable::handle
calls.
As a first step, this change replaces the Box that was used for storing
Widget objects so far with an Rc. In the future, this Rc will take over
the job that the outer Option is currently providing, which is to
decouple the Widget object from the Ui object itself. Doing so is
necessary, because we would like to provide a mutable Ui reference (in
the form of a MutCap) to the widget's handle method, but with the widget
being part of the Ui this will necessarily result in a borrowing
conflict. Having an Rc allows us to "clone" easily and cheaply and
decouple the two objects for the invocation at hand.