Remove Option around stored Widget objects
This change finally eliminates the need for "taking" any Widget objects
out of the Ui when invoking event handling methods. Along with that, it
makes the code more robust, because it removes potential panic paths
that could happen whenever a widget was taken already (and that we tried
to make unreachable by clever [but programmatically non-assertable] API
design).
Specifically, the patch removes the Option wrapping the Widget objects
internally in the Ui and with it the need for the Ui::with method. We
now just clone the Rc that we introduced earlier in order to eliminate
the need for taking a reference to a Widget while having a mutable
reference to the Ui.