Store hooks in reference counted manner
With upcoming changes we would like to give event hooks greater
flexibility in what they can accomplish: we want to provide them with a
MutCap instead of only a Cap. This will bring event hook functions on
par with regular event handlers in terms of capabilities, which will
simplify our logic and allow for a more coherent design with even
further reaching changes that will introduce async/await support for
dealing with events.
To that end, this change stores the indices for hooked widgets in a
reference counted container. In so doing, we can support the case where
widgets remove their hooks from the even hook handler itself. Our
inability of dealing properly with this corner case was the main reason
for not providing hook functions with a MutCap to begin with.