Make all event handling async
Now that we have separated widgets from their mutable data and event
handling happens on non-mutable object references, we can tackle the
main reason for all these refactorings: we want to move event handling
to a much nicer async/await based programming model.
This change marks a first step in this effort. Specifically, it converts
all event handling methods on the Ui and the Handleable trait to async
ones. Unfortunately, "true" async methods in traits are not yet fully
supported by Rust and so we need to fall back to using the async-trait
crate. Using this crate we get the desired functionality, but we incur a
heap allocation for every future returned. That is considered
acceptable, because 1) for now performance is simply not a big issue for
existing use cases and 2) this will eventually get fixed with minimal
effort on our side once the Rust compiler supports async methods in
traits.