Parametrize Ui and Widget with new message type
One of the problems we identified with the existing design is that we
tried to have a single mechanism -- event delivery -- to address two
use cases. In a nutshell, a UI in our sense should be able to receive
events from the outside and react to them. A "reaction" can take various
forms, but all but very simple components (widgets) may want to
communicate with other components in the system.
So far we have piggy-backed on the existing event delivery mechanism for
this purpose. However, that arguably conflates the two concerns.
Instead, we propose a separate mechanism for sending messages within the
UI, without involvement of any kind of event. While events typically
"bubble up" (or down, depending on one's view), messages are intended as
a more directed way for communication, where unhandled ones are not just
propagated to parent widgets.
Treading on this path, this change introduces a new generic parameter
(in addition to the existing one for the event type) to the UI itself
and the Handleable trait (as well as various traits also involved). This
parameter identifies the message type to use. Upcoming changes will make
use of this message type to implement a new communication mechanism.