Introduce message sending mechanism
This change makes use of the message type, by which we parametrized
various UI components earlier, to implement a new communication
mechanism. To that end, we extend the MutCap trait with a new async
method, send, that can be used to send a message to the widget with the
provided ID. Furthermore, we enhance the Handleable trait with a new
async handler method, react, that marks the endpoint for messages sent
this way.