Introduce "call" support
Unfortunately, in some scenarios sending a message by transferring
ownership of a bit of data may not be feasible. Originally, this thought
caused us to introduce the UiEvent::Returnable variant, which provided
a notion of communication without transfer of ownership of the data
involved.
This change now introduces a similar communication mechanism that is
async. Specifically, we add a new method, call, to the MutCap trait and
a corresponding async handler, respond, to Handleable. Contrary to send,
call works on a mutable reference to a message, making sure that the
sender stays the owner, while still allowing receivers to modify the
data if required. We still allow for return of a message as well, as
there are cases where a completely different message will ultimately be
returned.